# webapp/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VooZuu App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/forgotPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: auth_ctrl
    post:
      tags: ['Auth Module']
      description: User Forgot Password
      summary: User Forgot Password
      # used as the method name of the controller
      operationId: forgotPassword
      parameters:
        - name: body
          in: body
          description: Forgot Password
          schema:
           $ref: '#/definitions/forgotPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/resetPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: auth_ctrl
    post:
      tags: ['Auth Module']
      description: Admin Reset Password
      summary: Admin Reset Password
      # used as the method name of the controller
      operationId: resetPassword
      parameters:
        - name: body
          in: body
          description: Reset Password
          schema:
           $ref: '#/definitions/resetPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/activation/{token}:
    # binds a127 app logic to a route
    x-swagger-router-controller: auth_ctrl
    get:
      tags: ['Auth Module']
      description: User Activation By Id
      summary: User Activation By Id
      # used as the method name of the controller
      operationId: activation
      parameters:
        - in: path
          name: token
          description: The user id
          required: true
          type: string              
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                         
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  forgotPassword:
    type: object
    properties:
      email:
        type: string
  resetPassword:
    type: object
    properties:
      resetKey:
        type: string
      password:
        type: string
...